change: calendar event date and time formatting
The formatting of dates and times for calendar events have been updated. Import of the intl package was necessary for this modification. A redundant comment block was also removed to clean up the code. Different ways were introduced to format the event time based on different conditions like if it's an all-day event or if the start and end time of the event are the same.